+++ settings/permissions/PermissionSets.js
@@ -16,34 +16,31 @@ import PermissionSetForm from './PermissionSetForm';
 class PermissionSets extends React.Component {
   static propTypes = {
     label: PropTypes.string.isRequired,
-    location: PropTypes.object.isRequired,
-    history: PropTypes.shape({
-      push: PropTypes.func.isRequired,
-    }).isRequired,
-    stripes: PropTypes.shape({
-      hasPerm: PropTypes.func.isRequired,
-    }).isRequired,
     resources: PropTypes.shape({
-      permissionSets: PropTypes.shape({
+      entries: PropTypes.shape({
         records: PropTypes.arrayOf(PropTypes.object),
       }),
     }).isRequired,
     mutator: PropTypes.shape({
       permissionSets: PropTypes.shape({
         POST: PropTypes.func,
+        PUT: PropTypes.func,
         DELETE: PropTypes.func,
       }),
     }).isRequired,
   };
 
   static manifest = Object.freeze({
-    permissionSets: {
+    entries: {
       type: 'okapi',
       records: 'permissions',
       DELETE: {
         path: 'perms/permissions',
       },
       POST: {
+        path: 'perms/permissions',
+      },
+      PUT: {
         path: 'perms/permissions',
       },
       GET: {
@@ -53,119 +50,23 @@ class PermissionSets extends React.Component {
     },
   });
 
-  constructor(props) {
-    super(props);
-
-    this.onAdd = this.onAdd.bind(this);
-    this.onEdit = this.onEdit.bind(this);
-    this.onSave = this.onSave.bind(this);
-    this.onRemove = this.onRemove.bind(this);
-    this.onCancel = this.onCancel.bind(this);
-
-    const path = props.location.pathname;
-    const selectedId = (/perms\//.test(path))
-      ? /perms\/(.*)$/.exec(path)[1]
-      : null;
-
-    this.state = { selectedId };
-    this.callout = null;
-  }
-
-  onAdd() {
-    this.showLayer('add');
-  }
-
-  onEdit(permSet) {
-    this.setState({ selectedId: permSet.id });
-    this.showLayer('edit');
-  }
-
-  onSave(permSet) {
-    const action = (permSet.id) ? 'PUT' : 'POST';
-    return this.props.mutator.permissionSets[action](permSet)
-      .then(() => this.hideLayer());
-  }
-
-  onRemove(permSet) {
-    return this.props.mutator.permissionSets.DELETE(permSet).then(() => {
-      this.showCalloutMessage(permSet.displayName);
-      this.hideLayer();
-    });
-  }
-
-  onCancel(e) {
-    e.preventDefault();
-    this.hideLayer();
-  }
-
-  showCalloutMessage(name) {
-    const message = (
-      <span>
-        The permission set <strong>{name || 'Untitled Permission Set' }</strong>
-        was successfully <strong>deleted</strong>
-      </span>
-    );
-
-    this.callout.sendCallout({ message });
-  }
-
-  showLayer(name) {
-    this.props.history.push(`${this.props.location.pathname}?layer=${name}`);
-  }
-
-  hideLayer() {
-    this.props.history.push(`${this.props.location.pathname}`);
-  }
-
   render() {
-    const { resources, location, stripes } = this.props;
-    const permissionSets = (resources.permissionSets || {}).records || [];
-    const query = location.search ? queryString.parse(location.search) : {};
-    const selectedItem = (query.layer === 'edit')
-      ? _.find(permissionSets, p => p.id === this.state.selectedId) : {};
-
-    const container = document.getElementById('ModuleContainer');
-    if (!container) return (<div />);
-
-    const addMenu = (
-      <IfPermission perm="perms.permissions.item.post">
-        {/* In practice, there is point letting someone create a set if they can't set its name */}
-        <IfPermission perm="perms.permissions.item.put">
-          <PaneMenu>
-            <Button title="Add Permission Set" onClick={this.onAdd} buttonStyle="primary paneHeaderNewButton">
-              + New
-            </Button>
-          </PaneMenu>
-        </IfPermission>
-      </IfPermission>
-    );
-
     return (
-      <EntrySelector
+      <EntryManager
         {...this.props}
-        onAdd={this.onAdd}
-        onEdit={this.onEdit}
-        detailComponent={PermissionSetDetails}
-        contentData={permissionSets}
         parentMutator={this.props.mutator}
+        entryList={_.sortBy((this.props.resources.entries || {}).records || [], ['displayName'])}
+        detailComponent={PermissionSetDetails}
         paneTitle={this.props.label}
-        detailPaneTitle="Permission Set Details"
+        entryLabel="permission set"
+        entryFormComponent={PermissionSetForm}
         nameKey="displayName"
-        paneWidth="70%"
-        addMenu={addMenu}
-      >
-        <Layer isOpen={!!(query.layer)} label="Edit Permission Set" container={container}>
-          <PermissionSetForm
-            initialValues={selectedItem}
-            stripes={stripes}
-            onSave={this.onSave}
-            onCancel={this.onCancel}
-            onRemove={this.onRemove}
-          />
-
-        </Layer>
-        <Callout ref={ref => (this.callout = ref)} />
-      </EntrySelector>
+        permissions={{
+          put: 'perms.permissions.item.post',
+          post: 'perms.permissions.item.post',
+          delete: 'perms.permissions.item.delete',
+        }}
+      />
     );
   }
 }