+++ test/ui-testing/patron_group.js
@@ -80,8 +80,8 @@ module.exports.test = function(uiTestCtx) {
         })
         .then(function(result) {
           userid = result
-          console.log('        (found user ID ' + userid + ")")
           done()
+          console.log('        (found user ID ' + userid + ")")
         })
         .catch(done)
       })
@@ -96,8 +96,8 @@ module.exports.test = function(uiTestCtx) {
         .xtract('id("adduser_group")/option[contains(.,"' + gid + '" )]/@value')
         .then(function(result) {
           communityid = result
-          console.log('        (found patron group ID ' + communityid + ")")
           done()
+          console.log('        (found patron group ID ' + communityid + ")")
         })
         .catch(done)
       })
@@ -112,14 +112,21 @@ module.exports.test = function(uiTestCtx) {
       })
       it('should fail at deleting "' + gid + '" group', done => {
         nightmare
-        .wait(2222)
+        .wait(1111)
         .click(config.select.settings)
         .wait(wait)
         .click('a[href="/settings/users"]')
         .wait('a[href="/settings/users/groups"]')
-        .wait(wait)
         .click('a[href="/settings/users/groups"]')
-	.wait(wait)
+	.wait(function(dp) {
+	  var dnode = document.evaluate(dp, document, null, XPathResult.FIRST_ORDERED_NODE_TYPE, null)
+	  if (dnode.singleNodeValue) {
+	    return true
+	  }
+	  else {
+	    return false
+	  }
+	},deletePath)
         .xclick(deletePath)
         .click('a[href="/settings/users/addresstypes"]')
         .wait(wait)
@@ -153,8 +160,8 @@ module.exports.test = function(uiTestCtx) {
         .xtract('id("adduser_group")/option[contains(.,"Staff")]/@value')
         .then(function(result) {
           staffid = result
-          console.log('        (found "Staff" group ID ' + staffid +")")
           done()
+          console.log('        (found "Staff" group ID ' + staffid +")")
         })
         .catch(done)
       })
@@ -168,13 +175,21 @@ module.exports.test = function(uiTestCtx) {
       })
       it('should delete "' + gid + '" patron group', done => {
         nightmare
-        .wait(wait)
+        .wait(1111)
         .xclick('//span[.="Settings"]')
         .wait(wait)
         .xclick('id("ModuleContainer")//a[.="Users"]')
         .wait(wait)
         .xclick('id("ModuleContainer")//a[.="Patron groups"]')
-        .wait(wait)
+	.wait(function(dp) {
+	  var dnode = document.evaluate(dp, document, null, XPathResult.FIRST_ORDERED_NODE_TYPE, null)
+	  if (dnode.singleNodeValue) {
+	    return true
+	  }
+	  else {
+	    return false
+	  }
+	},deletePath)
         .xclick(deletePath)
         .wait(parseInt(process.env.FOLIO_UI_DEBUG) ? parseInt(config.debug_sleep) : 555) // debugging
         .then(result => { done() })
@@ -182,6 +197,7 @@ module.exports.test = function(uiTestCtx) {
       }) 
       it('should confirm that "' + gid + '" patron group has been deleted', done => {
         nightmare
+	.wait(wait)
         .evaluate(function(gid) {
           var cnode = document.evaluate('//div[.="' + gid + '"]', document, null, XPathResult.FIRST_ORDERED_NODE_TYPE, null)
           if (cnode.singleNodeValue) {