+++ CHANGELOG.md
@@ -37,6 +37,13 @@
 * Refactor permission sets. Fulfills UIU-231, UIU-276, and UIU-165.
 * Change search label. Fixes UIU-140.
 * Sort addresses by primary field. Fixes UIU-296.
+* Refactor permission sets to use `<EntryManager>`. Fixes UIU-300.
+* Refactor async user validator. Fixes UIU-302.
+* All network communication is now done through stripes-connect, and `fetch` is unused. Fixes UIU-304.
+* Break out core search/sort functionality into re-usable `<SearchAndSort>` component. Fixes UIU-278.
+* Generalise `<SearchAndSort>` by moving all user-specific functionality out of it. Fixes UIU-299.
+* Rewire proxy selection with the new `proxiesfor` endpoint. Fixes UIU-292.
+* Add `optionalDependencies` on find-user plugin. Fixes UIU-313.
 
 ## [2.10.1](https://github.com/folio-org/ui-users/tree/v2.10.1) (2017-09-05)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v2.10.0...v2.10.1)